Finishing implementation
Completed the test_app.py file by adding in the tokens into the tests.
Minor changes in app.py to get the test file to work (uncommented 'return app')
Created a .env file as well as added dotenv implementation so that the environment variables can be accessed correctly.
Added a references.txt file.